# services/payments-svc/internal/server/openapi.yaml
...
openapi: 3.0.3
info:
    title: Omnibus network payment hub API
    version: 1.0.0
paths:
...
    /v1/operator/clock:
        post:
            tags:
                - NetworkOperationsService
            description: |-
                SetClock moves the scenario clock (and the chain's) forward. Demo builds
                 only; a production build answers PERMISSION_DENIED.
            operationId: NetworkOperationsService_SetClock
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.SetClockRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.SetClockResponse'
...
    /v1/ops/defunds:
        post:
            tags:
                - BankOperationsService
            description: RequestDefund asks for free position back to the master account (maker).
            operationId: BankOperationsService_RequestDefund
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.RequestDefundRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Defund'
    /v1/ops/defunds/{id}/approve:
        post:
            tags:
                - BankOperationsService
            description: ApproveDefund is the checker; never the requester, always the same bank.
            operationId: BankOperationsService_ApproveDefund
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.ApproveDefundRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Defund'
...
components:
    schemas:
...
        api.payments.v1.ApproveDefundRequest:
            type: object
            properties:
                id:
                    type: string
...
        api.payments.v1.Defund:
            type: object
            properties:
                id:
                    type: string
                amount:
                    $ref: '#/components/schemas/api.payments.v1.Money'
                requestedBy:
                    type: string
                approvedBy:
                    type: string
                status:
                    type: string
                createdAt:
                    type: string
                    format: date-time
...
        api.payments.v1.Money:
            type: object
            properties:
                amount:
                    type: string
                currency:
                    type: string
            description: |-
                Money is an amount in a currency. The amount is a decimal string with at
                 most two fraction digits ("25000000.00"), never a float.
...
        api.payments.v1.RequestDefundRequest:
            type: object
            properties:
                amount:
                    $ref: '#/components/schemas/api.payments.v1.Money'
...
        api.payments.v1.SetClockRequest:
            type: object
            properties:
                time:
                    type: string
                    format: date-time
        api.payments.v1.SetClockResponse:
            type: object
            properties:
                now:
                    type: string
                    format: date-time
...